fix: file-worker任务状态更新请求无序到达导致第三方源文件偶现分发失败 #2434
1.修复Trace数据断链；
2.优化重调度逻辑，支持参数可配置；
3.重构部分代码。
