+++ src/backend/job-file-gateway/service-job-file-gateway/src/main/java/com/tencent/bk/job/file_gateway/dao/filesource/FileSourceTaskDAO.java
@@ -39,6 +39,8 @@ public interface FileSourceTaskDAO {
 
     FileSourceTaskDTO getFileSourceTaskById(String id);
 
+    FileSourceTaskDTO getFileSourceTaskByIdForUpdate(String id);
+
     Long countFileSourceTasksByBatchTaskId(String batchTaskId, Byte status);
 
     List<FileSourceTaskDTO> listByBatchTaskId(String batchTaskId);
