+++ src/backend/job-file-gateway/service-job-file-gateway/src/main/java/com/tencent/bk/job/file_gateway/dao/filesource/FileTaskDAO.java
@@ -38,6 +38,8 @@ public interface FileTaskDAO {
 
     int deleteFileTaskByFileSourceTaskId(String fileSourceTaskId);
 
+    FileTaskDTO getFileTaskByIdForUpdate(Long id);
+
     FileTaskDTO getOneFileTask(String fileSourceTaskId, String filePath);
 
     List<FileTaskDTO> listFileTasks(String fileSourceTaskId, Integer start, Integer pageSize);
