+++ src/backend/job-file-gateway/service-job-file-gateway/src/main/java/com/tencent/bk/job/file_gateway/dao/filesource/impl/FileTaskDAOImpl.java
@@ -135,6 +135,20 @@ public class FileTaskDAOImpl extends BaseDAOImpl implements FileTaskDAO {
     }
 
     @Override
+    public FileTaskDTO getFileTaskByIdForUpdate(Long id) {
+        List<Condition> conditions = new ArrayList<>();
+        conditions.add(defaultTable.ID.eq(id));
+        val record = dslContext.selectFrom(defaultTable).where(
+            conditions
+        ).forUpdate().fetchOne();
+        if (record == null) {
+            return null;
+        } else {
+            return convertRecordToDto(record);
+        }
+    }
+
+    @Override
     public FileTaskDTO getOneFileTask(String fileSourceTaskId, String filePath) {
         List<Condition> conditions = new ArrayList<>();
         conditions.add(defaultTable.FILE_SOURCE_TASK_ID.eq(fileSourceTaskId));
