+++ src/backend/job-file-gateway/service-job-file-gateway/src/main/java/com/tencent/bk/job/file_gateway/service/impl/FileSourceTaskServiceImpl.java
@@ -27,38 +27,28 @@ package com.tencent.bk.job.file_gateway.service.impl;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.tencent.bk.job.common.constant.ErrorCode;
 import com.tencent.bk.job.common.exception.InternalException;
-import com.tencent.bk.job.common.exception.NotFoundException;
 import com.tencent.bk.job.common.exception.ServiceException;
 import com.tencent.bk.job.common.model.Response;
 import com.tencent.bk.job.common.model.http.HttpReq;
-import com.tencent.bk.job.common.util.ArrayUtil;
-import com.tencent.bk.job.common.util.file.FileSizeUtil;
-import com.tencent.bk.job.common.util.file.PathUtil;
 import com.tencent.bk.job.common.util.http.JobHttpClient;
 import com.tencent.bk.job.common.util.json.JsonUtils;
-import com.tencent.bk.job.execute.common.constants.FileDistStatusEnum;
 import com.tencent.bk.job.file_gateway.consts.TaskCommandEnum;
 import com.tencent.bk.job.file_gateway.consts.TaskStatusEnum;
-import com.tencent.bk.job.file_gateway.dao.filesource.FileSourceBatchTaskDAO;
 import com.tencent.bk.job.file_gateway.dao.filesource.FileSourceDAO;
 import com.tencent.bk.job.file_gateway.dao.filesource.FileSourceTaskDAO;
 import com.tencent.bk.job.file_gateway.dao.filesource.FileTaskDAO;
 import com.tencent.bk.job.file_gateway.dao.filesource.FileWorkerDAO;
-import com.tencent.bk.job.file_gateway.model.dto.FileSourceBatchTaskDTO;
 import com.tencent.bk.job.file_gateway.model.dto.FileSourceDTO;
 import com.tencent.bk.job.file_gateway.model.dto.FileSourceTaskDTO;
 import com.tencent.bk.job.file_gateway.model.dto.FileTaskDTO;
 import com.tencent.bk.job.file_gateway.model.dto.FileTaskProgressDTO;
 import com.tencent.bk.job.file_gateway.model.dto.FileWorkerDTO;
-import com.tencent.bk.job.file_gateway.model.resp.inner.FileLogPieceDTO;
 import com.tencent.bk.job.file_gateway.model.resp.inner.FileSourceTaskStatusDTO;
 import com.tencent.bk.job.file_gateway.model.resp.inner.TaskInfoDTO;
 import com.tencent.bk.job.file_gateway.model.resp.inner.ThirdFileSourceTaskLogDTO;
 import com.tencent.bk.job.file_gateway.service.DispatchService;
 import com.tencent.bk.job.file_gateway.service.FileSourceTaskService;
-import com.tencent.bk.job.file_gateway.service.context.TaskContext;
-import com.tencent.bk.job.file_gateway.service.context.impl.DefaultTaskContext;
-import com.tencent.bk.job.file_gateway.service.listener.FileTaskStatusChangeListener;
+import com.tencent.bk.job.file_gateway.service.FileSourceTaskUpdateService;
 import com.tencent.bk.job.file_gateway.service.remote.FileSourceTaskReqGenService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.helpers.MessageFormatter;
@@ -70,7 +60,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -82,7 +71,7 @@ public class FileSourceTaskServiceImpl implements FileSourceTaskService {
 
     public static final String PREFIX_REDIS_TASK_LOG = "job:file-gateway:taskLog:";
 
-    private final FileSourceBatchTaskDAO fileSourceBatchTaskDAO;
+    private final FileSourceTaskUpdateService fileSourceTaskUpdateService;
     private final FileSourceTaskDAO fileSourceTaskDAO;
     private final FileTaskDAO fileTaskDAO;
     private final FileWorkerDAO fileworkerDAO;
@@ -91,21 +80,18 @@ public class FileSourceTaskServiceImpl implements FileSourceTaskService {
     private final FileSourceTaskReqGenService fileSourceTaskReqGenService;
     private final RedisTemplate<String, Object> redisTemplate;
     private final JobHttpClient jobHttpClient;
-    private final List<FileTaskStatusChangeListener> fileTaskStatusChangeListenerList = new ArrayList<>();
 
     @Autowired
-    public FileSourceTaskServiceImpl(
-        FileSourceBatchTaskDAO fileSourceBatchTaskDAO,
-        FileSourceTaskDAO fileSourceTaskDAO,
-        FileTaskDAO fileTaskDAO,
-        FileWorkerDAO fileworkerDAO,
-        FileSourceDAO fileSourceDAO,
-        DispatchService dispatchService,
-        FileSourceTaskReqGenService fileSourceTaskReqGenService,
-        @Qualifier("jsonRedisTemplate") RedisTemplate<String, Object> redisTemplate,
-        FileTaskStatusChangeListener fileTaskStatusChangeListener,
-        JobHttpClient jobHttpClient) {
-        this.fileSourceBatchTaskDAO = fileSourceBatchTaskDAO;
+    public FileSourceTaskServiceImpl(FileSourceTaskUpdateService fileSourceTaskUpdateService,
+                                     FileSourceTaskDAO fileSourceTaskDAO,
+                                     FileTaskDAO fileTaskDAO,
+                                     FileWorkerDAO fileworkerDAO,
+                                     FileSourceDAO fileSourceDAO,
+                                     DispatchService dispatchService,
+                                     FileSourceTaskReqGenService fileSourceTaskReqGenService,
+                                     @Qualifier("jsonRedisTemplate") RedisTemplate<String, Object> redisTemplate,
+                                     JobHttpClient jobHttpClient) {
+        this.fileSourceTaskUpdateService = fileSourceTaskUpdateService;
         this.fileSourceTaskDAO = fileSourceTaskDAO;
         this.fileTaskDAO = fileTaskDAO;
         this.fileworkerDAO = fileworkerDAO;
@@ -114,11 +100,6 @@ public class FileSourceTaskServiceImpl implements FileSourceTaskService {
         this.fileSourceTaskReqGenService = fileSourceTaskReqGenService;
         this.redisTemplate = redisTemplate;
         this.jobHttpClient = jobHttpClient;
-        addFileTaskStatusChangeListener(fileTaskStatusChangeListener);
-    }
-
-    public void addFileTaskStatusChangeListener(FileTaskStatusChangeListener listener) {
-        this.fileTaskStatusChangeListenerList.add(listener);
     }
 
     @Override
@@ -197,65 +178,6 @@ public class FileSourceTaskServiceImpl implements FileSourceTaskService {
             fileWorkerDTO.getCloudAreaId(), fileWorkerDTO.getInnerIp());
     }
 
-    private void writeLog(FileSourceTaskDTO fileSourceTaskDTO,
-                          FileWorkerDTO fileWorkerDTO,
-                          FileTaskProgressDTO fileTaskProgressDTO) {
-        String taskId = fileSourceTaskDTO.getId();
-        String fileSizeStr = FileSizeUtil.getFileSizeStr(fileTaskProgressDTO.getFileSize());
-        ThirdFileSourceTaskLogDTO thirdFileSourceTaskLog = new ThirdFileSourceTaskLogDTO();
-        String sourceCloudIp = fileWorkerDTO.getCloudIp();
-        thirdFileSourceTaskLog.setIp(sourceCloudIp);
-        // 追加文件源名称
-        // 日志定位坐标：（文件源，文件路径），需要区分不同文件源下相同文件路径的日志
-        FileSourceDTO fileSourceDTO = fileSourceDAO.getFileSourceById(fileSourceTaskDTO.getFileSourceId());
-        if (fileSourceDTO == null) {
-            throw new NotFoundException(ErrorCode.FILE_SOURCE_NOT_EXIST,
-                ArrayUtil.toArray("fileSourceId:" + fileSourceTaskDTO.getFileSourceId()));
-        }
-        String filePathWithSourceAlias = PathUtil.joinFilePath(
-            fileSourceDTO.getAlias(),
-            fileTaskProgressDTO.getFilePath()
-        );
-        List<FileLogPieceDTO> fileLogPieceList = new ArrayList<>();
-        FileLogPieceDTO fileLogPiece = new FileLogPieceDTO();
-        fileLogPiece.setContent(buildFileLogContent(fileTaskProgressDTO, filePathWithSourceAlias, fileSizeStr));
-        fileLogPiece.setDisplaySrcFile(filePathWithSourceAlias);
-        fileLogPiece.setProcess(buildProcessStr(fileTaskProgressDTO));
-        fileLogPiece.setSize(fileSizeStr);
-        fileLogPiece.setSrcIp(sourceCloudIp);
-        fileLogPiece.setStatus(FileDistStatusEnum.PULLING.getValue());
-        fileLogPiece.setStatusDesc(FileDistStatusEnum.PULLING.getName());
-        fileLogPieceList.add(fileLogPiece);
-        thirdFileSourceTaskLog.setFileTaskLogs(fileLogPieceList);
-        // 写入Redis
-        redisTemplate.opsForList().rightPush(PREFIX_REDIS_TASK_LOG + taskId, thirdFileSourceTaskLog);
-        // 一小时后过期
-        redisTemplate.expireAt(PREFIX_REDIS_TASK_LOG + taskId, new Date(System.currentTimeMillis() + 3600 * 1000));
-    }
-
-    @SuppressWarnings("StringBufferReplaceableByString")
-    private String buildFileLogContent(FileTaskProgressDTO fileTaskProgressDTO,
-                                       String filePathWithSourceAlias,
-                                       String fileSizeStr) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("FileName: ");
-        sb.append(filePathWithSourceAlias);
-        sb.append(" FileSize: ");
-        sb.append(fileSizeStr);
-        sb.append(" ");
-        sb.append("Speed: ");
-        sb.append(fileTaskProgressDTO.getSpeed());
-        sb.append(" Progress: ");
-        sb.append(fileTaskProgressDTO.getProgress());
-        sb.append("% Detail: ");
-        sb.append(fileTaskProgressDTO.getContent());
-        return sb.toString();
-    }
-
-    private String buildProcessStr(FileTaskProgressDTO fileTaskProgressDTO) {
-        return "" + fileTaskProgressDTO.getProgress() + "%";
-    }
-
     @Override
     public String updateFileSourceTask(FileTaskProgressDTO fileTaskProgressDTO) {
         String fileSourceTaskId = fileTaskProgressDTO.getFileSourceTaskId();
@@ -272,97 +194,19 @@ public class FileSourceTaskServiceImpl implements FileSourceTaskService {
             );
             return null;
         }
-        TaskStatusEnum previousStatus = TaskStatusEnum.valueOf(fileTaskDTO.getStatus());
-        fileTaskDTO.setDownloadPath(fileTaskProgressDTO.getDownloadPath());
         FileSourceTaskDTO fileSourceTaskDTO = fileSourceTaskDAO.getFileSourceTaskById(fileSourceTaskId);
         if (fileSourceTaskDTO == null) {
             log.error("Cannot find fileSourceTaskDTO by taskId {} filePath {}", fileSourceTaskId, filePath);
             return null;
         }
-        return updateFileSourceTask(fileTaskDTO, fileSourceTaskDTO, fileTaskProgressDTO, previousStatus);
-    }
-
-    @Transactional(value = "jobFileGatewayTransactionManager", rollbackFor = {Throwable.class})
-    public String updateFileSourceTask(FileTaskDTO fileTaskDTO,
-                                       FileSourceTaskDTO fileSourceTaskDTO,
-                                       FileTaskProgressDTO fileTaskProgressDTO,
-                                       TaskStatusEnum previousStatus) {
-        // 开启事务后立即加排它锁，保证读取到其他事务已提交的数据
-        FileSourceBatchTaskDTO fileSourceBatchTaskDTO =
-            fileSourceBatchTaskDAO.getBatchTaskByIdForUpdate(fileSourceTaskDTO.getBatchTaskId());
-        String fileSourceTaskId = fileTaskProgressDTO.getFileSourceTaskId();
-        String filePath = fileTaskProgressDTO.getFilePath();
-        TaskStatusEnum status = fileTaskProgressDTO.getStatus();
-        Integer progress = fileTaskProgressDTO.getProgress();
-        Long fileSize = fileTaskProgressDTO.getFileSize();
-        if (fileSourceBatchTaskDTO == null) {
-            log.error("Cannot find fileSourceBatchTaskDTO by batchTaskId {} fileSourceTaskId {} filePath {}",
-                fileSourceTaskDTO.getBatchTaskId(),
-                fileSourceTaskId,
-                filePath
-            );
-            return null;
-        }
-        FileWorkerDTO fileWorkerDTO = fileworkerDAO.getFileWorkerById(fileSourceTaskDTO.getFileWorkerId());
-        int affectedRowNum = -1;
-        if (status == TaskStatusEnum.RUNNING) {
-            // 已处于结束态的任务不再接受状态更新
-            if (!fileTaskDTO.isDone()) {
-                fileTaskDTO.setProgress(progress);
-                fileTaskDTO.setFileSize(fileSize);
-                fileTaskDTO.setStatus(TaskStatusEnum.RUNNING.getStatus());
-                affectedRowNum = fileTaskDAO.updateFileTask(fileTaskDTO);
-                logUpdatedTaskStatus(fileSourceTaskId, filePath, progress, status);
-            } else {
-                log.info("fileTask {} already done, do not update to running", fileSourceTaskId);
-            }
-        } else if (status == TaskStatusEnum.SUCCESS) {
-            fileTaskDTO.setProgress(100);
-            fileTaskDTO.setStatus(TaskStatusEnum.SUCCESS.getStatus());
-            affectedRowNum = fileTaskDAO.updateFileTask(fileTaskDTO);
-            logUpdatedTaskStatus(fileSourceTaskId, filePath, progress, status);
-        } else if (status == TaskStatusEnum.FAILED) {
-            fileTaskDTO.setProgress(progress);
-            fileTaskDTO.setStatus(TaskStatusEnum.FAILED.getStatus());
-            affectedRowNum = fileTaskDAO.updateFileTask(fileTaskDTO);
-            logUpdatedTaskStatus(fileSourceTaskId, filePath, progress, status);
-        } else if (status == TaskStatusEnum.STOPPED) {
-            fileTaskDTO.setProgress(progress);
-            fileTaskDTO.setStatus(TaskStatusEnum.STOPPED.getStatus());
-            affectedRowNum = fileTaskDAO.updateFileTask(fileTaskDTO);
-            logUpdatedTaskStatus(fileSourceTaskId, filePath, progress, status);
-        } else {
-            log.warn("fileTask {} unknown status:{}", fileSourceTaskId, status);
-        }
-        if (affectedRowNum != -1) {
-            log.info("{} updated, affectedRowNum={}", fileTaskDTO, affectedRowNum);
-        }
-        // 通知关注者
-        if (status != previousStatus) {
-            notifyFileTaskStatusChangeListeners(fileTaskDTO, fileSourceTaskDTO, fileWorkerDTO, previousStatus, status);
-        }
-        // 进度上报
-        writeLog(fileSourceTaskDTO, fileWorkerDTO, fileTaskProgressDTO);
-        return fileSourceTaskId;
-    }
-
-    private void notifyFileTaskStatusChangeListeners(FileTaskDTO fileTaskDTO,
-                                                     FileSourceTaskDTO fileSourceTaskDTO,
-                                                     FileWorkerDTO fileWorkerDTO,
-                                                     TaskStatusEnum previousStatus,
-                                                     TaskStatusEnum currentStatus) {
-        TaskContext context = new DefaultTaskContext(fileTaskDTO, fileSourceTaskDTO, fileWorkerDTO);
-        if (!fileTaskStatusChangeListenerList.isEmpty()) {
-            boolean stop;
-            for (FileTaskStatusChangeListener listener : fileTaskStatusChangeListenerList) {
-                stop = listener.onStatusChange(context, previousStatus, currentStatus);
-                if (stop) break;
-            }
-        }
-    }
-
-    private void logUpdatedTaskStatus(String taskId, String filePath, Integer progress, TaskStatusEnum status) {
-        log.info("updated fileTask:{},{},{},{}", taskId, filePath, progress, status.name());
+        String batchTaskId = fileSourceTaskDTO.getBatchTaskId();
+        Long fileTaskId = fileTaskDTO.getId();
+        return fileSourceTaskUpdateService.updateFileSourceTask(
+            batchTaskId,
+            fileSourceTaskId,
+            fileTaskId,
+            fileTaskProgressDTO
+        );
     }
 
     @Override
