+++ src/backend/job-file-worker-sdk/service-job-file-worker-sdk/src/main/java/com/tencent/bk/job/file/worker/task/heartbeat/HeartBeatTask.java
@@ -30,9 +30,9 @@ import com.tencent.bk.job.common.util.http.JobHttpClient;
 import com.tencent.bk.job.common.util.json.JsonUtils;
 import com.tencent.bk.job.common.util.machine.MachineUtil;
 import com.tencent.bk.job.file.worker.config.WorkerConfig;
-import com.tencent.bk.job.file.worker.cos.service.EnvironmentService;
-import com.tencent.bk.job.file.worker.cos.service.GatewayInfoService;
-import com.tencent.bk.job.file.worker.cos.service.MetaDataService;
+import com.tencent.bk.job.file.worker.service.EnvironmentService;
+import com.tencent.bk.job.file.worker.service.GatewayInfoService;
+import com.tencent.bk.job.file.worker.service.MetaDataService;
 import com.tencent.bk.job.file_gateway.model.req.inner.HeartBeatReq;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.tuple.Pair;
