+++ src/backend/job-file-worker/service-job-file-worker/src/main/java/com/tencent/bk/job/file/worker/cos/impl/COSFileResourceImpl.java
@@ -35,15 +35,15 @@ import com.tencent.bk.job.file.worker.api.IFileResource;
 import com.tencent.bk.job.file.worker.cos.JobTencentInnerCOSClient;
 import com.tencent.bk.job.file.worker.cos.consts.COSActionCodeEnum;
 import com.tencent.bk.job.file.worker.cos.consts.COSNodeTypeEnum;
-import com.tencent.bk.job.file.worker.cos.service.COSBaseService;
-import com.tencent.bk.job.file.worker.cos.service.COSRemoteClient;
-import com.tencent.bk.job.file.worker.cos.service.MetaDataService;
-import com.tencent.bk.job.file.worker.cos.service.RemoteClient;
 import com.tencent.bk.job.file.worker.model.BucketDTO;
 import com.tencent.bk.job.file.worker.model.FileDTO;
 import com.tencent.bk.job.file.worker.model.req.BaseReq;
 import com.tencent.bk.job.file.worker.model.req.ExecuteActionReq;
 import com.tencent.bk.job.file.worker.model.req.ListFileNodeReq;
+import com.tencent.bk.job.file.worker.service.COSBaseService;
+import com.tencent.bk.job.file.worker.service.COSRemoteClient;
+import com.tencent.bk.job.file.worker.service.MetaDataService;
+import com.tencent.bk.job.file.worker.service.RemoteClient;
 import com.tencent.bk.job.file_gateway.model.resp.common.FileNodesDTO;
 import com.tencent.bk.job.file_gateway.model.resp.common.FileTreeNodeDef;
 import com.tencent.cos.model.Bucket;
